Move embed related code to separate cog
